+++ packages/react-app/app/widget/page.tsx
@@ -1,0 +1,10 @@
+"use client";
+import { Widget } from '@/components/Widget';
+
+export default function MiniAppWidget() {
+  return (
+    <main>
+      <Widget />
+    </main>
+  );
+}
