+++ packages/react-app/components/Layout.tsx
@@ -1,54 +1,7 @@
 import { FC, ReactNode, useEffect, useState } from "react";
 import Header from "./Header";
-import { Drawer, Button, Typography, styled } from "@mui/material";
+import { Drawer, Button } from "@mui/material";
 import { useWidgetEvents, WidgetEvent } from "@bandohq/widget";
-import Image from "next/image";
-import Carousel from "react-multi-carousel";
-import "react-multi-carousel/lib/styles.css";
-
-const CarouselDiv = styled("div")({
-  "& .carousel-container": {
-    margin: "0 0 0 20px",
-  },
-  "& .carousel-item": {
-    height: "220px",
-    padding: "0 5px",
-  },
-  "& .custom-dot-list-style": {
-    position: "relative",
-  },
-  "& .custom-dot-list-style .react-multi-carousel-dot button": {
-    margin: "0 0.5rem",
-    border: "none",
-    backgroundColor: "#9A9A9A",
-    width: "10px",
-    height: "10px",
-  },
-  "& .custom-dot-list-style .react-multi-carousel-dot.react-multi-carousel-dot--active button": {
-    backgroundColor: "#1A1A1A",
-  },
-});
-
-const responsiveCarousel = {
-  desktop: {
-    breakpoint: { max: 3000, min: 1024 },
-    items: 2,
-    slidesToSlide: 1, // optional, default to 1.
-    partialVisibilityGutter: 140
-  },
-  tablet: {
-    breakpoint: { max: 1024, min: 464 },
-    items: 2,
-    slidesToSlide: 1, // optional, default to 1.
-    partialVisibilityGutter: 140
-  },
-  mobile: {
-    breakpoint: { max: 464, min: 0 },
-    items: 1,
-    slidesToSlide: 1, // optional, default to 1.
-    partialVisibilityGutter: 140
-  }
-};
 
 interface Props {
   children: ReactNode;
@@ -82,58 +35,6 @@ const Layout: FC<Props> = ({ children }) => {
     <>
       <div className="bg-gypsum overflow-hidden flex flex-col min-h-screen">
         <Header />
-        <div className="flex flex-shrink-0 items-center">
-          <Image
-            style={{ width: "100%", height: "auto", padding: "0 20px" }}
-            className="block h-8 sm:block w-auto lg:block"
-            src="/mini_app_banner.svg"
-            width="100"
-            height="20"
-            alt="Mini App Banner"
-          />
-        </div>
-        <div>
-          <Typography variant="h6" sx={{ p: '0 20px', m: '20px 0'}}>🔥 Hottest Deals</Typography>
-        </div>
-        <CarouselDiv>
-          <Carousel
-            swipeable={true}
-            draggable={false}
-            arrows={false}
-            showDots={true}
-            responsive={responsiveCarousel}
-            ssr={true} // means to render carousel on server-side.
-            infinite={true}
-            autoPlay={true}
-            autoPlaySpeed={5000}
-            renderDotsOutside
-            keyBoardControl={true}
-            customTransition="all 1s linear"
-            transitionDuration={700}
-            partialVisible={true}
-            containerClass="carousel-container"
-            removeArrowOnDeviceType={["tablet", "mobile"]}
-            dotListClass="custom-dot-list-style"
-            itemClass="carousel-item"
-          >
-            <img
-              style={{ height: "220px" }}
-              className="block h-8 sm:block w-auto lg:block"
-              src="/example_deal_1.svg"
-              width="100"
-              height="20"
-              alt="Mini App Banner"
-            />
-            <img
-              style={{ height: "220px" }}
-              className="block h-8 sm:block w-auto lg:block"
-              src="/example_deal_2.svg"
-              width="100"
-              height="20"
-              alt="Mini App Banner"
-            />
-          </Carousel>
-        </CarouselDiv>
         <div className="max-w-7xl mx-auto" style={{ width: "100%" }}>
           {children}
         </div>
